Significantly refactored `AuthController.cs` to inherit from `BaseControllerApi`, injecting new services like `ILogger`, `EmailService`, and `UserManager`. Updated login and registration methods to use new request models and added endpoints for email confirmation.

Created `BaseControllerApi.cs` for common API functionality. Updated `CartController` and `UsersController` to inherit from the new base class. Enhanced `OrdersController` with a new `CreateOrder` method and updated `ProductsController` to include product reviews and variants.

Updated `DependencyInjection.cs` to register new services and repositories. Refactored `AuthService`, added `JwtTokenGenerator` and `RedisTokenService` for token management, and introduced `EmailService` for email functionalities.

Created `UserService` for user-related operations and updated repository interfaces. Modified `SeedData.cs` to align with new entity structures.

Updated data models in `Order.cs`, `OrderDetail.cs`, and `Payment.cs`, and refactored product-related files for consistency.

Updated `PetDatabaseContext.cs` and migration files to reflect schema changes. Implemented `RedisCartRepository` for efficient cart management.

Overall, improved application structure for better maintainability and readability.

+++ FurEverCarePlatform.Domain/Entities/Product.cs
@@ -9,7 +9,7 @@ public class Product : BaseEntity
     public Guid StoreId { get; set; }
     public string Name { get; set; }
     public string? Description { get; set; }
-    public decimal BasePrice { get; set; }
+    public float BasePrice { get; set; }
     public decimal Weight { get; set; }
     public decimal Length { get; set; }
     public decimal Height { get; set; }
@@ -61,7 +61,7 @@ public class ProductVariant : BaseEntity
 {
     public Guid ProductId { get; set; }
     public JsonDocument Attributes { get; set; }
-    public decimal Price { get; set; }
+    public float Price { get; set; }
     public int Stock { get; set; }
 
     // Navigation property nếu dùng EF Core
